[FEAT] 비동기로 combine처리하는 requsetAPI메서드 구현(#65)
