+++ Plu-iOS/Plu-iOS/Scenes/Onboarding/OnboardingInteface/NicknameCheck.swift
@@ -22,13 +22,13 @@ protocol NicknameCheck {
                                    to checker: textFieldVaildChecker) -> textFieldOutput
     func getNicknameVaildPublisher(from checker: textFieldVaildChecker,
                                    with manager: NicknameManager) -> textFieldOutput
-    func makeNicknameResultPublisher(from input: textFieldInput,
+    func nicknamePublisher(from input: textFieldInput,
                                      to checker: textFieldVaildChecker,
                                      with manager: NicknameManager) -> textFieldOutput
 }
 
 extension NicknameCheck where Self: AnyObject {
-    func makeNicknameResultPublisher(from input: textFieldInput, to checker: textFieldVaildChecker, with manager: NicknameManager) -> textFieldOutput {
+    func nicknamePublisher(from input: textFieldInput, to checker: textFieldVaildChecker, with manager: NicknameManager) -> textFieldOutput {
         let stateFromNicknamePublisher = self.getNicknameStatePublisher(from: input, to: checker)
         let nickNameValidPublisher = self.getNicknameVaildPublisher(from: checker, with: manager)
         return stateFromNicknamePublisher.merge(with: nickNameValidPublisher).eraseToAnyPublisher()
